+++ modules/bot/components/crud/rules/ExplodeStringFieldComponent.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace app\modules\bot\components\rules;
+namespace app\modules\bot\components\crud\rules;
 
 /**
  * Class ExplodeStringFieldComponent
