+++ modules/bot/components/crud/rules/LocationToArrayFieldComponent.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace app\modules\bot\components\rules;
+namespace app\modules\bot\components\crud\rules;
 
 /**
  * Class LocationToArrayFieldComponent
