+++ modules/bot/components/crud/services/AttributeButtonsService.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace app\modules\bot\services;
+namespace app\modules\bot\components\crud\services;
 
 use app\components\helpers\ArrayHelper;
 use app\modules\bot\components\Controller;
