+++ modules/bot/components/crud/services/ViewFileService.php
@@ -1,7 +1,7 @@
 <?php
 
 
-namespace app\modules\bot\services;
+namespace app\modules\bot\components\crud\services;
 
 use app\modules\bot\components\Controller;
